pyproject.toml
[tool.poetry]
name = "damage-prediction-ae"
version = "0.1.0"
description = ""
authors = ["Yacine Bel-Hadj <yacine.bel-hadj@vub.be>" , "Francisco de Nolasco Santos <>"]
readme = "README.md"
packages = [ 
    { include = "src" }
]
[tool.poetry.dependencies]
python = "^3.10"
notebook = "^7.2.2"
dynaconf = "^3.2.6"
pyrootutils = "^1.0.4"
torchdata = "^0.9.0"
dask = "^2024.11.2"
numpy = "^2.1.3"
pandas = "^2.2.3"
fastparquet = "^2024.11.0"
torchaudio = "^2.5.1"
matplotlib = "^3.9.3"
pyarrow = "^18.1.0"
sqlalchemy = "^2.0.36"
api-24sea = "^1.4.12"
requests = "^2.32.3"
pytorch-lightning = "^2.4.0"
mlflow = "^2.18.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
